+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   dictionary when necessary
 * [HexaPDF::Importer#import] to always return an imported object with the same
   class as the argument
+* [HexaPDF::Type::OutlineItem] to always be an indirect object
 
 
 ## 0.26.2 - 2022-10-22